refactor: extract DOM utilities into lib/dom-utils.ts factory
Extract 7 inline DOM utility functions (waitForPageTransition,
isElementVisible, asHTMLElement, setInputValue, fireMouseEvent,
activateElement, findVueParentByName) from content.ts into
createDomUtils(deps) DI factory at lib/dom-utils.ts.

Reduces content.ts by 133 lines. Follows established factory pattern
with typed DomUtilsDeps interface and injected win/doc/getPaginationInfo.
